+++ patterns/rotatingRainbowHue.h
@@ -1,4 +1,4 @@
-void rotatingRainbowHue(uint16_t numLED) {
+void rotatingRainbowHue(uint8_t numLED) {
   static uint8_t rotatingRainbowStartHue = 0;
 
   fill_rainbow(leds, numLED, rotatingRainbowStartHue, 7);
